Fix 0-width window bug by removing window size hints.
You used to be able to resiz the image smaller than 1:1 by setting
the window's geometry hints to min_width and height of 1, but if you
wapped the mouse to the left, you could get a 0-width window which
would hang on 100% CPU (doing what???) and not repainting if you enlarged
it again. You also lost the "File" menu heading and WM's buttons.
Without the geometry hints, you can only enlarge the image, not reduce it,
and with a 1x1 image the smallest size leaves the menu heading.