Image grows and shrinks but menu takes whole window
Before, the list of menu items covered the File toolbar item, were normal
size but the image could not be resized. Now we've put the toolbar and
image in the vbox and got the scaling right but if you open the menu
the items cover the whole image, blotting it out.