Avoid GTK scaler bug by doing resizes to 1xH and Wx1 in two steps
https://bugzilla.gnome.org/show_bug.cgi?id=80925
Image scaler slows down and uses huge amounts of RAM for high-ratio
reductions. Avoid the worst cases (scaling to width of height of 1) by
checking for those cases and scaling each axis separately.
This improves matters but still exhibits the bad behaviour, albeit less
so, when scaling to small sizes.